+++ src/main/java/step3/RacingGamePlayer.java
@@ -1,5 +1,7 @@
 package step3;
 
+import step3.view.RacingInputView;
+
 public class RacingGamePlayer {
 
     public void playRacingGame() {
